requirements.txt
# Polymath v4 Dependencies

# Core
psycopg2-binary>=2.9.9
numpy>=1.26.0

# PDF Parsing
pymupdf>=1.23.0

# Embeddings (local GPU)
sentence-transformers>=2.2.0
torch>=2.1.0

# Google Cloud (Gemini)
google-cloud-aiplatform>=1.40.0
google-cloud-storage>=2.14.0

# Neo4j
neo4j>=5.15.0

# Web/API
requests>=2.31.0
huggingface-hub>=0.20.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
